+++ Applanga.xcframework/macos-arm64_x86_64/ApplangaOSX.framework/Versions/Current/Headers/ApplangaOSX.h
@@ -1,0 +1,8 @@
+//
+//  ApplangaOSX.h
+//  Applanga
+//
+//  Created by Steffen Römer on 23/12/15.
+//  Copyright © 2015 Mbaas Development GmbH. All rights reserved.
+//
+#import <ApplangaOSX/Applanga.h>