+++ ApplangaWatchOS.framework/Headers/ApplangaWatchOS.h
@@ -1,0 +1,20 @@
+//
+//  ApplangaWatchOS.h
+//  ApplangaWatchOS
+//
+//  Created by Richard Elms on 30.10.20.
+//  Copyright © 2020 Mbaas Development GmbH. All rights reserved.
+//
+
+#import <Foundation/Foundation.h>
+#import <ApplangaWatchOS/Applanga.h>
+
+//! Project version number for ApplangaWatchOS.
+FOUNDATION_EXPORT double ApplangaWatchOSVersionNumber;
+
+//! Project version string for ApplangaWatchOS.
+FOUNDATION_EXPORT const unsigned char ApplangaWatchOSVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <ApplangaWatchOS/PublicHeader.h>
+
+