+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 *Applanga iOS Documentation:* <https://www.applanga.com/docs-integration/ios> 
 ***
 
+### Version 2.0.144 (6 Jan 2021)
+#### Added
+- Support for watchOS and MacOS
+- Draft menu redesign
+
+---
+
 ### Version 2.0.140 (24 Sep 2020)
 #### Added
 - Error when key is too long